Update download_toggle_video2.py
Implements a fix as suggested in #49. Should address #47 and #51.

ffmpeg argument is now `-user_agent` instead of `-user-agent`.

Removed the `-headers 'origin: http://video.toggle.sg/\r\n'` string, which was originally added in response to #44. Not sure why it has since broken the script, but this could be due to using an old version of ffmpeg. 

Also includes some more debugging statements to capture the environment (OS and python versions).